+++ lib/Crypto/SelfTest/PublicKey/test_import_RSA.py
@@ -389,11 +389,6 @@ Lr7UkvEtFrRhDDKMtuIIq19FrL4pUIMymPMSLBn3hJLe30Dw48GQM4UCAwEAAQ==
         self.assertEqual(key.n, self.n)
         self.assertEqual(key.e, self.e)
 
-    def test_export_key(self):
-        key = RSA.construct([self.n, self.e, self.d, self.p, self.q, self.pInv])
-        der = key.export_key("DER")
-        self.assertEqual(der, self.rsaKeyDER)
-
 
 class ImportKeyFromX509Cert(unittest.TestCase):